Day10/2 ready
I finally realized it's just a linear equation system.
So I was doing some research what is the best linear
system solver for positive integers and based
google + wikipedia +chatgpt + gemini, it was google's ortools
with its CBC solver.

Of course it had some quirks around it's source with a
self-calling cyclic target-dependency which meant that I could
not make it work with my mingw-clang toolchain, and I was
also unable to build it with visual studio.

So after 6 hours of suffering I changed course and moved
to MSVC temporarily because there is a prebuilt binary
version of ortools which links with MSVC runtimes,
so I was finally in.

After that I figured out with the help of gemini
how to solve an equation system with ortools.
and tadadaaa. Result.